+++ test/service_test.rb
@@ -28,9 +28,13 @@ class ServiceTest < Minitest::Test
 
     FileUtils.rm_rf(TMP_DIR)
     FileUtils.mkdir_p(TMP_DIR)
+
+    Potluck.config = Potluck::Config.new(dir: TMP_DIR)
   end
 
   def teardown
+    Potluck.config = Potluck::Config.new
+
     FileUtils.rm_rf(TMP_DIR)
   end